Fix search for unicode patterns
The fts4 index now uses the unicode61 tokenizer instead of 'simple' which
only supports ASCII characters [1].

Still, this failed for japanese substrings, e.g. when searching for
"大好き" in "おすし大好き！".  As a workaround the search falls back
to a LIKE query if there are unicode runes in the search pattern from
the letter, other category [2].

[1]: https://www.sqlite.org/fts3.html#unicode61
[2]: https://golang.org/pkg/unicode/#Lo