Add relesae job
This job reuses the artifacts that were compiled by the build job.
It took me quite a while to set this up since the download-artifact actions documentation is lacking.
I was able to figure out the downloaded artifact paths by adding a step that recursively prints the working directory directory

- name: ls recursively
  run: ls -R